CMS-4316 Add manualOrderValue to index
Added as double for now, we have to see if it should be indexed as long, but this requires some thinking in relation to other numeric values

+++ modules/wem-core/src/main/java/com/enonic/wem/core/entity/SetNodeChildOrderCommand.java
@@ -65,7 +65,7 @@ public class SetNodeChildOrderCommand
 
             final Node node = doGetNode( nodeIdOrderValue.getNodeId(), false );
 
-            final Node editedNode = Node.editNode( node ).manualOrderValue( nodeIdOrderValue.getManualOrderValue() ).build();
+            final Node editedNode = Node.editNode( node ).manualOrderValue( (long) nodeIdOrderValue.getManualOrderValue() ).build();
 
             doStoreNode( editedNode );
         }
